Refactor `CustomChip` widget to be more generic for extensibility.
The `CustomActionChip` widget is created to extend the CustomChip
much like the Flutter provided `Chip` widget is extended by the
`ActionChip`

The `CategoryChip` is used to toggle the color of the chip widget
and will later be used to send the value as well to the form data
model. This specificity of features meant that a custom widget was
necessary.